+++ test/dummy/lib/dummy/engine.rb
@@ -1,0 +1,5 @@
+module Dummy
+  class Engine < ::Rails::Engine
+    isolate_namespace Dummy
+  end
+end